Allow some options to specified as URL parameters.
If server and player are specified (i.e. as
`index.html?server=example.org&player=Dewin`),
try to auto-connect to the specified server and player slot.
A password can be specified, if required, using password=PASSWORD.

?hideui=1 will completely hide the entire UI including commmand input.
This is intended for capturing the spectator client as a browser source
inside OBS.